Migrated StreamDecoder to an expect/actual
